pyproject.toml
[project]
name = "ya-reviews-mcp"
version = "0.1.0"
description = "MCP server for Yandex Maps business reviews"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Ruben Khachaturov" }
]
keywords = ["mcp", "yandex", "maps", "reviews", "business"]
dependencies = [
    "fastmcp>=2.13.0,<2.15.0",
    "playwright>=1.40.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
    "click>=8.1.0",
    "uvicorn>=0.27.1",
    "starlette>=0.40.0",
]

[project.urls]
Repository = "https://github.com/mrkhachaturov/ya-reviews-mcp"

[project.scripts]
ya-reviews-mcp = "ya_reviews_mcp:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ya_reviews_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
